+++ js/main.js
@@ -1,5 +1,4 @@
 import {showScreen} from './utils';
 import greetingScreenContent from './greeting/greeting';
 
-
 showScreen(greetingScreenContent);
