--- js/result/result.js
@@ -1,9 +1,0 @@
-import template from './result-view';
-import {getElementFromTemplate, showScreen} from '../utils';
-import greetingScreenContent from '../greeting/greeting';
-
-const resultContent = getElementFromTemplate(template);
-const buttonReplay = resultContent.querySelector(`.main-replay`);
-buttonReplay.addEventListener(`click`, () => showScreen(greetingScreenContent));
-
-export default resultContent;
